finagle: Remove Toggle type and override `isDefinedAt`
Problem
1. Toggle look ups allocate a lot of `java.lang.Integer` due to auto boxing when
calling `Toggle.apply(int)`, this is possibly because `Toggle[T]` extends
`PartialFunction[Int, Boolean]` and scala is not using the specialized apply
function.

2. Some `Toggle.isDefinedAt` is calling `PartialFunctions.isDefinedAt`, but the
call triggers auto-boxing because the `pf.isDefinedAt` is not specialized. And
in our case, all `PartialFunctions` we passed to create a `Toggle` always returns
true when calling its `isDefinedAt`.

Solution
1. Remove the type in Toggle and set `isDefinedAt` as true by default.

2. Set `isDefinedAt` default to be true in `Toggle` abstract class, and
remove the call to `pf.isDefinedAt` if that is guaranteed to return true.

JIRA Issues: CSL-9538

Differential Revision: https://phabricator.twitter.biz/D456960