Move parse phase into its own machine.
This allows issuing this step independently and then re-use the result multiple times.
This hides a lot of the underlying protocol, so it's not clear it will stay this way. Also, the extended query machine would benefit from having the `Flush` removed.
This could be done by allowing a Send command that allows multiple messages.