finagle-core: Requeue/RetryFilter tracing annos
Problem:
MethodBuilder will tack on a RetryFilter for application level retries. The
sampled traces should have some visibility in what's going on here as in
D459106.

Solution / Result:
Mimic the requeue annotations in the RetryFilter. For requeued/retried requests,
we set the rpc name as (requeue|retry).

JIRA Issues: CSL-9606

Differential Revision: https://phabricator.twitter.biz/D466083