Adds basic rich client.
Basic support for a rich client. Allows prepared statements and one-off statements.
Named prepared statements are assumed to be cached per connection (this is what finagle-mysql does).
The rationale is that it allows releasing the connection after preparing the statement which can then be reused for executing other statements.
This is important since the time between preparing a statement and executing it can be large, so reserving a connection simply to wait for execution is not reasonable.
The downside is that the prepare step can be executed multiple times (increasing latency) as well as using more resources on the postgresql backend.