Filter IPv6 address-es when NIC has only local IPv6 addresses
Netty fixed that bug at
https://github.com/netty/netty/pull/10170

That bug is still present in at least openjdk (1.8.0_191, 11.0.1).

More precisely `InetAddress.getAllByName`
sends both AAAA and A queries for a given host.

When a DNS server returns IPv4/IPv6 address-es and
that the IPv6 address is chosen by the client,
subsequent queries will fail, given that network of
the NIC is not able to route IPv6 packets.