finagle-partitioning: Add a PartitionNodeManager
Problem

To support the custom partitioning strategy, we need to have a different node
manager from the HashRingNodeManager to map from a partition id to a service
node.

Solution

Introduce this PartitionNodeManager listens to destination changes to alter the
service node, and stores service nodes in a map structure.

Result

A service node has a set of Address of a logical partition's all instances, the
default FailureAcrrualFactory will be used for each endpoint, however, it will
be disabled heuristically when the instance number is not greater than 3.

Differential Revision: https://phabricator.twitter.biz/D465868