finagle-stats: Don't compute the histogram bucket while holding the lock
Problem

In the BucketedHistogram we acquire the lock before computing the
bucket index which can be expensive if it gets into the logarithm
function. However, we don't need to synchronize the computation.

Solution

Move the computation of the bucket index to before lock acquisition and
only hold the lock while writing values.

Differential Revision: https://phabricator.twitter.biz/D488812