finagle/thrift: Add scala custom partitioning
Problem

Introduce the custom partitioning service for Thrift/ThriftMux scala client.

Solution

Custom Partitioning maps requests directly to partitions by specify the
partition id. partitionId can be a physical host identifier or a logical id that
represents a group of hosts. This is set by a reverse lookup
`getLogicalPartition`.

When a client is specified with the CustomPartitioningStrategy, the client
stack materializes the `ThriftCustomPartitioningService` with a
`ParititonNodeManager`.

JIRA Issues: CSL-9217

Differential Revision: https://phabricator.twitter.biz/D487174