util|scrooge|finagle|twitter-server|finatra: Use `skip in publish`
Problem

sbt introduced `skip in publish` here: https://github.com/sbt/sbt/issues/3136,
which replaces the need to configure multiple different settings in attempt
to skip publishing of a sub-project within a multi-module project. The
current workarounds to skip publishing of a sub-project appear to lack some
efficacy in practice.

Solution

Switch to using `skip in publish` and also configure local overwrite
to true to aid in release publishing.

Differential Revision: https://phabricator.twitter.biz/D491635