finagle-thrift: Suppress stack trace generation in the UncaughtAppExceptionFilter
Problem
When an uncaught exception is caught by UncaughtAppExceptionFilter, it creates
a TApplicationException just to serialise and send it to the client. The stack
trace is apparently meaningless and not serialised, however it is populated, which
causes unnecessary memory allocations and consumes precious cpu cycles.

Solution
Suppress stack trace generation.

Differential Revision: https://phabricator.twitter.biz/D487942