finagle-core: Introduce BalancingPool
Problem

Pipelined (and multiplexed) protocols may incur head-of-line blocking (e.g.
from the server's processing threads or the network) when operating over a single
session. In certain cases it makes sense to replicate the number of sessions to
a given endpoint and load balance over them to avoid this. For example, we offer
this behavior in Finagle's memcached by default.

The current implementation of this operates at the load balancer layer in the stack
and replicates addresses as distinct endpoint stacks in the load balancer. This results
in replicating all the endpoint modules per each session! For example, we mint a fresh
`FailureAccrual` instance for each of these sessions and they each have to independently
converge on `Status.Closed` state to trip the circuit-breaker for a particular endpoint.

Solution

Introduce a new type of pool, `BalancingPool`, which composes a `SingletonPool` with
latency sensitive load balancing (i.e. `p2cPeakEwma`). This is a module which should be
installed at the pooling layer of the endpoint stack.

Result

Per-endpoint modules now avoid replication when replicating sessions.

JIRA Issues: CSL-9640

Differential Revision: https://phabricator.twitter.biz/D480376