finagle-zipkin-core: Remove flush and late-arrival annotations
Problem
-------

The annotations that are written into a span when it is forcibly flushed from
the DeadlineSpanMap can have timestamps up to 2 minutes later than the true
end of the trace. This extends the apparent overall duration of the trace,
which makes it harder to view in the UI and distorts the request duration
distributions for aggregate analysis.

Solution
--------

Remove the annotations.

Result
------

Trace end times are not artificially extended.

JIRA Issues: csl-9863

Differential Revision: https://phabricator.twitter.biz/D498073