finagle-core: disable FailFast when a client has one endpoint
Problem

There are lots of knobs and best practices to understand when using
Finagle.  For example, it is a recommended best practice to disable
`FailFastFactory` when a client is connected to a single endpoint.
This is because `FailFastFactory` fails closed which has two
implications: (1) the underlying exception masked under a
FailFastException and (2) the client can't do anything meaningful
by breaking the circuit early since it doesn't have any other paths
in the load balancer.

Solution

Disable `FailFastFactory` by default when we resolve a destination
with a single endpoint.

Result

Less things to worry about and configure as a Finagle user.

JIRA Issues: CSL-6260

Differential Revision: https://phabricator.twitter.biz/D498911