finagle-core: annotate namer traces at the request level
Problem:
Namer annotations are added as a ServiceFactory module. Traces should be
request based and not connection based

Solution / Result:
Create these annotations as a Service filter and not ServiceFactory. As a
result of tracing on the request path, the `namer.success` and `namer.failure`
annotations were removed.

JIRA Issues: CSL-9831

Differential Revision: https://phabricator.twitter.biz/D492443