finagle-core: offload filter annotations
Problem:
The OffloadFilter was placed at the very top of the client stack. This means
it's client annotations were not placed under the corresponding child request's
generated span, causing duplicates in the parent.

Solution:
Move the TraceInitializer to the top of the client stack.

Result:
While we are here lets update the scaladoc in StackClient to match the modules
present.

JIRA Issues: CSL-9884

Differential Revision: https://phabricator.twitter.biz/D502521