finagle-zipkin-scribe: configure the scribe client to use the NullTracer
Problem:
The zipkin-scribe client is tracing itself occasionally

The native thrift client isn’t configured with the NullTracer. However, the
TracelessFilter is applied before it reaches the client, removing the current
trace id and tracers from the context. But, the client stack still has the
TraceInitializer filter which creates a new trace id and reads in the
DefaultTracer. Because there’s not a parent (cleared trace id earlier), it
creates an unsamped trace id. So that trace is sampled at the `initialSampleRate`

Solution / Result:
Configure the Thrift client with the NullTracer

JIRA Issues: CSL-9906

Differential Revision: https://phabricator.twitter.biz/D507318