finagle-serversets: Fix failing unit tests caused by broken zookeeper client.
**Problem**

These finagle-serversets unit tests are deterministically failing:
```
Zk2ResolverTest, ZkSessionEndToEndTest, ZkAnnouncerTest, ZkResolverTest
```
Making improvements or adding new features to finagle-serversets is not possible without solid unit tests unless one wants to directly test in production.

The tests are broken in different ways:
* Improperly constructed in process zookeeper client which blocks forever on connecting to the in process zookeeper server.
* Some tests have logical issues that require either a deprecation or rewrite.
* Some tests are flaky because of timing issue (i.e. with zookeeper session).

The scope of this patch is focus on the first problem - the improperly constructed zookeeper client. Other issues will be addressed in subsequent patches for the separation of concerns.

In existing code we use
```
InetSocketAddress(port number)
```
to construct a socket address and pass this to ZooKeeper client. By definition, this constructor will use wildcard address, which is ok for server side socket binding but not proper for client side because wildcard address is meaningless in the context of client side where it's not routable or reachable. This will lead to ZooKeeper client not be able to connect to the in process ZooKeeper server, with
```
java.net.ConnectException: Connection refused
```
. Not able to connect to the in process ZooKeeper server will lead to lots of downstream tests fail, such as the resolver tests will always return Pending resolve status for a name.

**Solution**

Instead of using the wildcard binding address for client, use the loopback address explicitly when constructing a socket address. This ensures the ZooKeeper client is properly constructed and can reach to the in process ZooKeeper server.

Because the in process ZooKeeper client and server is the foundation of many other tests, we need make sure both are in good shape. So add a dedicated unit test for them as well. Had this in place it would be much faster to identify the root cause of existing test failures.

**Result**

With this fix, all finagle-serversets unit tests are passed, except one test in ZKAnnouncerTest (flaky) and ZkSessionEndToEndTest (require rewrite). These tests (along with removal of the SKIP_CI tag) will be fixed in future patches.

JIRA Issues: COORD-13822, CSL-9374

Differential Revision: https://phabricator.twitter.biz/D516264