finagle-doc: Update Quickstart Project to Work with JDK11
Problem

The Finagle Quickstart SBT project does not work with JDK11. This
is due to some options in the 'sbt' file used as the wrapper for
SBT.

Solution

Let's remove the options that no longer work with JDK11 as well
as the deprecated ones, and bump the version of 2.12 used to
match the version in Finagle's build.sbt file.

Result

The Finagle Quickstart project should work with at least JDK
versions 8-11.

JIRA Issues: CSL-9945

Differential Revision: https://phabricator.twitter.biz/D517570