finagle-serversets: Fix one broken ZkAnnouncerTest unit test "announce a primary endpoint with metadata and additional endpoints".
**Problem**

The unit test "announce a primary endpoint with metadata and additional endpoints" is deterministically failing. It fails because the syntax of a valid path with the presence of additional end points that can be bind to zk2 resolver is in the format of
```
zk2!<host:port>!<path>!<endpoint>
```
, instead of

```
zk2!<host:port>!<path>:<endpoint>
```
that the unit test is using. As a result, zk2 resolver fail to resolve the path to a name.

**Solution**

Use the right path syntax.

**Result**

All unit test of ZkAnnouncerTest is happy now and passing.

JIRA Issues: COORD-13822

Differential Revision: https://phabricator.twitter.biz/D518023