finagle-netty4: Deal with Race in SslClientVerificationHandler
Problem

We've seen a scenario occur within the `SslClientVerificationHandler` where
`onHandshakeComplete.setDone` is called, and the `promise` chained in the
`respond` combinator was already failed. This causes the call to `promise.setSuccess`
to throw and end up propagating to the root monitor, which is undesirable.

Solution

Switch the call for `promise.setSuccess` to `promise.trySuccess`, so that it doesn't bomb
out. And if `trySuccess` fails, log out the state of the already satisfied `promise`.

JIRA Issues: CSL-9971

Differential Revision: https://phabricator.twitter.biz/D520818