finagle-serversets: Fix broken unit test ZkSessionEndToEndTest.
**Problem**

All tests in ZkSessionEndToEndTest are broken.

First problem is both tests will block indefinitely. The cause is the tests are using MockTimer but without ticking the timer, so the RetryStream will never get scheduled to execute, leading to reconnect after session expiration never happen. This is a regression introduced in a previous commit where MockTimer was introduced.

Second problem is in test "ZkSession.retrying", the expectation of sessions sequence was wrong due to not take the initial emitted value into consideration:  For a Var, the current value of this Var is emitted synchronously upon subscription, so the sessions checks are always off by one.

**Solution**

For the first problem, there are three ways to fix it.

* Correctly use a MockTimer by tick it. However, I can't get this work, so I decide to leave this as an exercise to the readers.

* Use zero back off, which will cause the Future used by RetryStream immediately satisfied. This is fine because the purpose of the back off is to avoid thundering herd which is not a concern in the context of our unit test; plus it has an added bonus of taking the Timer schedule out of equation, leading to more deterministic test.

* Use a real timer, such as ScheduledThreadPoolTimer. I've verified this works with both test. Comparing to the zero back off solution, this might be less deterministic due to scheduling glitches (if ever).

**Result**

ZkSessionEndToEndTest is happy and pass. And ALL finagle serversets unit tests are happy and pass.

JIRA Issues: COORD-13822

Differential Revision: https://phabricator.twitter.biz/D519786