finagle-memcached: Change Expiry Logging Level to Trace
Problem

If a service is sending a command with a negative expiry, there's a
good chance that they're doing so many many times over. Therefore, a log
level of 'warning' is too heavyweight for a production service.

Solution

Turn the log level down to 'trace' so that if a service is setting a
negative expiry, they aren't impacted in production, but do have the
ability to track down where the offending call is coming from.

JIRA Issues: CSL-9665

Differential Revision: https://phabricator.twitter.biz/D521834