finagle-http: Only add tracing annotations when streaming messages finish
Problem

For large or infinite payloads adding an annotation for every chunk
read in a HTTP request or response can cause a very large number of
records to pile up which can cause trouble with very little value.

Solution

Only add a record at the end of the stream.

JIRA Issues: CSL-9921

Differential Revision: https://phabricator.twitter.biz/D522543