finagle-core: Filter.andThenIf That Doesn't Require a Tuple
Problem

The current version of `Filter.andThenIf` requires the user to pass
both the conditional and the conditionally applied `Filter` together
as a tuple. The ergonomics of this style of method matched a prior
version of how Scala code used to be written at Twitter. However, given
the current set of standards, this ends up feeling clunky, and requiring
additional unnecessary tuple allocations.

Solution

Provide a variant of the `Filter.andThenIf` method which takes the
individual components, the conditional and the conditionally applied
`Filter`, as individual parameters.

Differential Revision: https://phabricator.twitter.biz/D523010