finagle-core: Add withFilter to allow MethodBuilder apply more configuration
Problem

There is no easy way to apply a per-endpoint configuration to MethodBuilder
other than Timeout, TraceInitializer, Retry. Specifically, a per-endpoint
partitioning strategy.

Solution

Introduce a `withFilter` in finagle-core MethodBuilder which can be later used
in thriftmux MethodBuilder to chain a Partitioning Filter. We keep the
partitioning logic in thriftmux and add the placeholder in core MethodBuilder
for now and restrict the visibility to Finagle.

Differential Revision: https://phabricator.twitter.biz/D522940