Create separate library in finagle/finagle-thrift to avoid tests depending on tests
Currently, there are tests in finagle/finagle-thriftmux that extend test classes in this target.
It would be more higienic if tests didn't depend on tests, so I created a `scala_library` with the same sources and dependencies that other tests can depend on.

No change should be observed, since every test dependee should be accounted for.

JIRA Issues: DPB-12609

Differential Revision: https://phabricator.twitter.biz/D524194