finagle-netty4: Add ability to limit the number of connections on a listener
Problem

Sometimes a server may want to limit the number of connections
it will accept from all clients, but we can't do that right now.

Solution

Add a way to drop excess connections early in their lifecycle.
Unfortunately there aren't API's available to simply ignore new
TCP streams (or reset them) before the client thinks there is a
connection so this is the best solution. Fortunately, most connections
are using protocols that do some sort of handshaking such as TLS
or cleartext mux handshakes so this isn't a real issue. Protocols
that could be affected are cleartext HTTP or vanilla thrift.

Differential Revision: https://phabricator.twitter.biz/D517737