finagle-thrift: Make partitioning APIs a little friendlier for Java
Problem

Partitioning APIs in Finagle are designed for Scala first, which can entail
a lot of boilerplate when using them from Java.

Solution

Ensure that it's easy to call the relevant partitioning methods from Java,
and add methods that are easier to call when it's not easy to do.

JIRA Issues: CSL-9216, CSL-9218

Differential Revision: https://phabricator.twitter.biz/D525770