finagle-zipkin-scribe: correctly scope the logical stats
Problem:
We add some logical stats on top of the native Thrift client since
it does not support method builder. However, the RetryFilter is
using the wrong StatsReceiver. Problem was not caught as there
weren't tests to verify the stats.

Solution:
Scope the logical stats correctly and add tests. Although scribe
should only return the `TryLater` and `Ok` enums, we should
handle the case were a non-existent enum is received.

Result:
While we are here, we can clean up some of this code. h/t @ccoco.

`RawZipkinTracer`'s statsReceiver argument is a no-op. Instead of referencing
`DefaultStatsReceiver` or `ClientStatsReceiver` singletons in the constructor
of the `ScribeRawZipinTracer` singleton which is unsafe (LoadService initialization),
we should make use of a passed in StatsReceiver.

JIRA Issues: CSL-10060

Differential Revision: https://phabricator.twitter.biz/D527531