finagle-thrift: Add MethodBuilder Partitioning Strategies
Problem

PartitioningStrategy APIs are mainly in PartialFunction format in order to
support a part of endpoints of a single Thrift service, the unused endpoints are
ignored. However, in ThriftMux, we have MethodBuilder to support per-endpoint
configuration, we could have a set of finer APIs to take functions, which are
determined to config one endpoint with specific Request/Response types.

Solution

Add MethodBuilder partitioning strategies (please see the diagram below).
Combine HashingStrategyException and PartitioningStrategyException.

JIRA Issues: CSL-9219

Differential Revision: https://phabricator.twitter.biz/D528788