finagle-thrift/scrooge: Ensure that partitioning works for Java clients
Problem

Although we have confidence in the mechanism underlying partitioning,
we built it by testing it against Scala code.  This means that we didn't
exercise it with Scrooge-generated code for Java, or when programmed from
Java.

Solution

Fix a bug in the Java-generated Scrooge code to align better with the
paritioning assumptions, and provide an end to end test in Java for
partitioning in ThriftMux.  While doing that, updated java end to end
tests to use Scrooge-generated code for Java.

JIRA Issues: CSL-9216, CSL-9218

Differential Revision: https://phabricator.twitter.biz/D526316