finagle-core: introduce experimental generic RoutingService
Problem

We have multiple implementations of Routing or Routes
throughout the Twitter-stack, primarily within Finagle, Finatra,
and Twitter-Server.

Solution

We start to build up a generic RoutingService and Route
definition that leverage util-routing, along with utilities for
generating them. We start off with some basic test examples
on how this would be utilized by different protocols/implementations.

This is experimental and locked down so that we can iterate
and make breaking changes in the open.

JIRA Issues: CSL-9618

Differential Revision: https://phabricator.twitter.biz/D529710