finagle-internal/mtls: export mux `OppTls` to registry
Problem

We need to expose the `OppTls` level in a way that can
be programmatically accessible at runtime.

Solution

Add the `OppTls` param to the module we use to export other
s2s/mtls specific data.

Result

We can now scrape a service's registry and inspect the `OppTls` levels.

JIRA Issues: CSL-9992

Differential Revision: https://phabricator.twitter.biz/D539624