finagle-http: introduce building blocks for OpenAPI v3.0 based URI Path
Problem/Solution

We want to introduce routing based on paths defined via the OpenAPI v3.0
Path spec standard. Let's introduce some of the building blocks for defining
a Path. We will follow up with the ability to parse and match against this
Path in future revisions.

JIRA Issues: CSL-9312, CSL-8944

Differential Revision: https://phabricator.twitter.biz/D530168