finagle-{partitioning, thrift}: Partitioning client supports dynamic resharding
Problem

The custom partitioning client doesn't support dynamic resharding, which means
that even some simple procedures like adding or removing instances from a
cluster isn't supported.  The reason we can't support dynamic resharding today
is that we split partitioning up into two phases, one where we partition the
request, and a separate one where we subdivide the cluster into partitions.  If
we change the way we're partitioning and only one of them has updated, we can
end up with an incoherent result.  One example of the kind of incoherency we can
see is that if a request is formulated for one partition, but that partition means
something different in the new scheme, we can end up with a result that doesn't
make sense in either the old scheme or the new scheme.

Imagine we have three partitions, A, B, C.  And we have 3 shards, 1, 2, 3.
Here's the mapping from partitions to shards in the original sharding scheme { A
=> 1, B => 2, C => 3}.  At the same time, requests are strings, and we use the
length of the string as the shard key.  str.length % 3 is the key, and here's
the mapping: { 0 => A, 1 => B, 2 => C }

As an example, "foo" has length 3, so it's sent to partition A.  That maps to
shard 1.  "barqu" has length 5, so it's sent to partition C.  That maps to shard
3.

Then we reshard.  In the new scheme, we still have three partitions, A, B, C.
But now they map to different shards, { A => 2, B => 3, C => 1}!  And the
requests map to different partitions too.  { 0 => B, 1 => C, 2 => A }.

So under the new scheme, "foo" goes to partition B, which goes to shard 3.
And "barqu" goes to partition A, which goes to shard 2.

However, there's a potential race here, where we may mix the two schemes. In
that case, "foo" can go to partition A (like in the first scheme) and then shard
2, like in the second scheme.  Similarly, "barqu" would go to partition C like
in the first scheme, but that would map to shard 1, like in the second scheme.

In both of those cases, the result is incoherent with either sharding scheme.

Beyond the race condition, if a function changes its behavior under the hood
without notifying Finagle, we won't be able to update the cached
`ServiceFactory`s until there's a serverset change.  There's currently no API to
retrigger partitioning.

Solution

We make a contract with customers who want to do resharding to provide Finagle
with pure functions to describe the sharding scheme, and also ask our customers
to fully describe the state that they need in order to reshard.  This means that
we can recompute the sharding scheme when either the user-provided state or the
serverset changes, and re-populate our ServiceFactory caches.  Further, when
that happens, we'll cache the functions that describe the sharding scheme for
the given state.  The cached functions are then used together as a pair, so that
we don't have the race conditions that we described before.

In order to do this, we provide two new ways of describing partitioning schemes,
one for when the user wants to supply their own state,
`ClientCustomStrategy.resharding`, and also one where the user wants access to
the remote serverset's current state, `ClientCustomStrategy.clusterResharding`.
To be consistent with the new methods, we changed the API for creating the
ClientCustomStrategy without resharding to `ClientCustomStrategy.noResharding`.

Result

Partition resharding is coherent, easy to use, and updates quickly.

JIRA Issues: CSL-9858

Differential Revision: https://phabricator.twitter.biz/D543466