finagle-core: remove RouteBuilder and RoutingServiceBuilder
Problem

`RouteBuilder` and `RoutingServiceBuilder` have moving parts that depend
on ordering to behave correctly. This leaves open potential for runtime
exceptions that could be caught at compile time.

Solution

Remove `RouteBuilder` as all parameters are required and the `Route` will
only be constructed within the finagle framework. Remove
`RoutingServiceBuilder` in favor of using `RouterBuilder` directly, which in
turn simplifies the generic type signature that we need to carry, so we can
remove the need for `UserReq`. Introduce `Route.transformed` as a
replacement for the functionality of the `RouteBuilder` and its transformed
result, so that we only need to be concerned with a
`RequestTransformingFilter` if the `RoutingService` absolutely needs it.

Result

Simpler generic type signatures, we are ensured that required items are
present when building a `Route` and `RoutingService` at compile time,
and we get to cut down on owned code as a result.

JIRA Issues: CSL-10111

Differential Revision: https://phabricator.twitter.biz/D543593