finagle: Reorder OffloadFilter.client in StackClient
Problem

In the registry for finagle clients the offload filter `OffloadWorkFromIO` does
not render because `RegistryEntryLifecycle` is before `OffloadFilter` in the
`StackClient.newStack` builder.

Solution

Reorder `RegistryEntryLifecycle` after `OffloadWorkFromIO`.

Result

The string representation of `FuturePool` enabled for `OffloadFilter` or
`disabled` if offload filter for the client is not enabled.

JIRA Issues: MOBDR-6296

Differential Revision: https://phabricator.twitter.biz/D548722