finagle: return a ServiceFactory in HashRingNodeManager
Problem
`finagle-redis` partitioning client uses `HashRingNodeManager` to return a
service for each node, this makes the node manager hold on to a single
connection and `FailureAccrualPolicy` is never kicked in to eject a node.
As a result, when a client disconnect to the service, new connection can
never be re-established.

Solution
Make `HashRingNodeManager` return a `ServiceFactory` instead of
`Future[Service]`

Result
`finagle-redis` client should be able to re-establish a connection after the
previous connection is closed.
This change also affects `finagle-memcached` client since it shares the same
`HashRingNodeManager`, and after the `FailureAccuralPolicy` kicks in, it will
throw `FailureAccrualException` instead of a `Failure` when a node is marked
as dead.

JIRA Issues: CSL-9819

Differential Revision: https://phabricator.twitter.biz/D495652