Adds postgres builtin types.
This adds an internal representation of postgres' type system.
It also generates types that are builtin to postgres, using one of its source files.

This is necessary to decode the data on the wire since the columns are identified by `oid`.
Clients have to be aware of what each `oid` is to be able to decode the values.

The type system also allows creating additional types (e.g.: an enum), but this isn't yet supported.