finagle-core: Refactor LoadBalancerFactory & TrafficDistributor to allow delegating endpoint management
Problem

We turn addresses into endpoints in the TrafficDistributor, which works well for
most paradigms.  However, it makes it impossible to share endpoints across load
balancers, or to move an endpoint from one load balancer to another.  Both of
these are valuable properties to have for partitioning, for schemes with shards
that may be in more than one partition simultaneously, and for re-sharding.

Solution

Update LoadBalancerFactory module to be responsible for endpoint management if
it's not already passed endpoints, and to pass through the endpoints if it's
given endpoints.  Also refactor some internal TrafficDistributor tooling so
that it's usable outside of TrafficDistributor, to allow the partitioning layer
to take over some of its responsibilities.

Result

No change in this patch.  Just a refactor for now.

JIRA Issues: CSL-10156

Differential Revision: https://phabricator.twitter.biz/D558729