finagle-redis: Add the RedisTracingFilter and RedisLoggingFilter back into the default Redis Stack
Problem

I was experimenting with the Tracing module and instrumented a Redis Client. I
noticed that Redis rpc calls did not get named unlike the Http calls that were
being traced. After some discussion on gitter, @mosesn pointed out that a commit
where the RedisLoggingFilter and RedisTracingFilter had been ported during the
move from Netty 3 -> Netty 4, but had not been added back to the stack.

Solution

I added the RedisLoggingFilter and the RedisTracingFilter to the default client stack.

Result

This will add stats and tracing back to the Redis Client by default.
I followed the mysql client for adding the role and module definitions. If there as a better
way to do this let me know and I can change it.

Signed-off-by: Jing Yan <jyan@twitter.com>

JIRA Issues: CSL-10150

Differential Revision: https://phabricator.twitter.biz/D558552