FIX: multi queries was sometimes skipping results.
The issue is pretty subtle due to the stateful nature of Reader.
One more reason to remove support for this.

Basically, the previous code was writing into the Reader, but failing to
record the write's pending state for the next write.

It's not cleat why this lead to skipping results as opposed to raising
an exception.