finagle-netty4: Introduce ExternalClientEngineFactory
Problem

The `Netty4ClientEngineFactory` and `Netty4ServerEngineFactory` are the
safest options in terms of correctness. That is the reason why they are
included as the default. However, they are not the best in terms of
performance, and we don't provide further examples of alternatives.

Solution

Move `ExternalClientEngineFactory` to the open source version of
Finagle, which shows a scenario where we create a few SSL contexts ahead
of time, effectively caching them, and then choose the correct one based
on the `SslClientConfiguration` options. This should provide some
further guidance for folks that are looking to create their own.

JIRA Issues: CSL-10310

Differential Revision: https://phabricator.twitter.biz/D572567