fix: specs2 no longer delegates to Diffable.
This changed recently in https://github.com/etorreborre/specs2/commit/59d0c3d0a8f22693b05fa50bfeb11a92c4b6c4c5

Because implementing a proper `equals` for `Json` would require pulling in a json library, we bring back the ability
to define a custom equals check in specs2 only and then use it where it matters.

The reason why this only matters for jsonb is because postgres will transform the input json (removing duplicate keys and normalizing numbers and such).
This results in a different json than the input, but still equivalent in the Json sense.