finagle-core: use NoStackTrace for unterminated stack exceptions
Problem

Finagle stacks generally start with a `nilStack`, which always fails with an exception. This exception captures a backtrace, which forces any classes in the call stack to remain in memory. Normally, this is not a problem, but if Finagle is used in a dynamic classloading environment such as OSGi or a Servlet container, especially with memoization, it can prevent old versions of an application from getting unloaded.

Solution

Use a shared exception object that extends `NoStackTrace` for `nilStack`.

Result

Unterminated stacks will fail with an exception that does not include a stack trace. Stacks will not include references to classes in the call stack that created them

Differential Revision: https://phabricator.twitter.biz/D572893