finagle-core: Remove incorrect explanation of ResponseClassifier in docs
Problem:

Documentation for RetryPolicy incorrectly states that a request will be retried
if a custom response classifier is provided. The ResponseClassifier only checks
whether or not to make a deposit to the retryBudget. It is not possible to tell
Finagle which application-level exceptions are retryable in the RequeueFilter.

Solution:

Remove this sentence.

JIRA Issues: CSL-10125

Differential Revision: https://phabricator.twitter.biz/D575036