finagle-core: Use the default Scala hashcode and equals for ServiceFactoryProxy
Problem
ServiceFactoryProxy extends scala/Proxy which calculates its hashcode
based on `self`, which is the ServiceFactory that the SFP is proxied to. This
results in an unstable hashcode when we use ServiceFactoryRef as an
implementation of ServiceFactoryProxy, because the SFR.update method will change
the reference of `self`, which will change the hashcode of the SFR/SFP.

Besides, SFP extends scala/Proxy which calculates its equals method based on not
only the reference of 2 SFP, but also the SF object that the SFP is proxied to.
This results in the following behavior:
val sf = new ServiceFactory
val sfp = new ServiceFactoryProxy(sf){}
sf == sfp // returns false
sfp == sf // returns true

Solution
Remove the dependency on Proxy, and use Scala default implementation to
calculate hashcode and equals method for SFP, which only replies on the SFP
reference.

Result
SFP should have a stable hashcode despite what SF it proxied to.
A SFP equals to another SFP if and only if their references equal, and vise versa.

Differential Revision: https://phabricator.twitter.biz/D569045