finagle: Handle floating point errors when biasing for load in daperture
Problem

When computing weights we often get values of 1.000000000002 or
0.999999999998 or similar numbers which are supposed to be identical but
are not due to floating point math errors. This causes the load based
bias to unduly favor nodes when they have equal non-zero weights.

Solution

If the weights are effectively the same, meaning being within 0.0001 of
each other, they weights are discounted and we just compare load.

JIRA Issues: CSL-10305

Differential Revision: https://phabricator.twitter.biz/D569968