fix: TLS spec in Github Actions
Docker in GH actions has a different behaviour than on MacOS.
The TLS spec needs to mount files in the container, but with specific ownership / permissions.
The different behaviour causes the mounted files to have different owner on MacOS than Linux, so we need to do additional work when running in CI (or probably any Linux host for that matter).