finagle-core: Be Defensive when Converting Streams to Strings
Problem

`RetryPolicy` converts a `Stream[Duration]` to a `String` in order
to have nice output for the representation of a `RetryPolicy`.
Unfortunately, we've run into instances where calling `toString` on a
`Stream` throws an `UnsupportedOperationException`. Given the stack
trace, this appears to be a Scala bug, and so we need to work around it.

Solution

Let's be defensive when calling `toString` on a `Stream[Duration]` inside of
`RetryPolicy`. By wrapping the call in a 'try/catch', we can return a
non-informative default value which only loses a little bit of
information.

JIRA Issues: CSL-10396

Differential Revision: https://phabricator.twitter.biz/D582199