refactor: change CommandTag model to avoid exceptions
The previous stategy had to throw an exception when the command tag didn't provide the number of affected rows.
This made it risky for client code to extract the number of affected rows withough checking the command tag explicitely.
The new model allows a simple pattern match to extract the number of rows:

```scala
commandTag match {
  case CommandTag.AffectedRows(_, rows) => rows
  case _ => 0
}
```

Arguably, the `rows` method could have returned something like `0`, but this is opinionated. This new modeling repreents the underlying protocol more closely.