finagle-thriftmux: Change use of getSimpleName to getName
Problem

In JDK 11, the behavior of Class#getSimpleName was corrected to return an empty
string when called on instances representing Scala anonymous classes. In JDK 8,
getSimpleName returns non-empty strings similar to "anon$1". If strings obtained from
calls to getSimpleName are required to be unique, this may cause undesired behavior
on newer versions of the JDK.

Solution

Remove the use of getSimpleName at line 266 in
finagle/finagle-thriftmux/src/main/scala/com/twitter/finagle/ThriftMux.scala and
replace it with getName.

JIRA Issues: CSL-10388

Differential Revision: https://phabricator.twitter.biz/D586406