finagle-thriftmux: eagerly establish stack when using method builder
Problem

The thriftmux mb stack uses the `DynamicPartitioningService` which
lifts stack creation to runtime. This is incompatible with features
like the lb eager connections which rely on eager stack materialization
to work correctly.

Solution

Initialize the underlying stack in `DynamicPartitioningService` for
non-partitioned requests when the instance is created.

Result

The thriftmux MethodBuilder now works with eager connections.

JIRA Issues: CSL-10430

Differential Revision: https://phabricator.twitter.biz/D589592