{finagle-partitioning,finagle-core}: Partitioning client now shares endpoint data between partitions & doesn't race
Problem

If an endpoint moves from one partition to another, we lose a ton of information
about it, like failure accrual info, and failfast.  Further, we currently
duplicate endpoints across partitions if it's in both partitions.

There's also a race condition that's possible with the current implementation.
Inflight requests during a reshard risk getting sent to the wrong server.  This
means that it's possible for a request to be sent to a server that it was never
appropriate for, both in the old scheme AND in the new scheme.  This comes from
relying on mutable state at the load-balancer layer, in the mutable Var that
keeps track of endpoints.

Solution

Pass endpoints directly to the traffic distributor, by doing endpoint creation
in the partitioning layer.  Use only immutable Vars in the load-balancer layer,
to avoid this race condition.

JIRA Issues: CSL-10156

Differential Revision: https://phabricator.twitter.biz/D548439