finagle-thriftmux: Fix infinite loop in failing downgrade negotiation
Problem

If our server mux session gets a close call but ends up completing the
handshake and subsequently failing the downgrade (likely due to tls
requirements) we can find ourselves in an infinite loop.

Solution

On the failure pathway don't consider the handshake done, but failed
(which is true) which exits the loop.

Differential Revision: https://phabricator.twitter.biz/D599296