Bumping zookeeper version due to security issues with Netty
Problem
The most recent release from finagle is using org.apache.zookeeper/zookeeper on
version 3.5.0-alpha.  This zookeeper version is using io.netty/netty on version
3.7.0 which introduces several vulnerable scenarios on the code.

Solution
The solution for this issue was bumping zkVersion on the project, to an
a compatible version of zookeeper that is using netty in a newer version (most
precisely, zookeeper was bumped to version 3.5.6).

Result
The result of this pull-request will be finagle be free from these vulnerable
scenarios from this outdated dependency. Closes #887.

Signed-off-by: Yufan Gong <yufang@twitter.com>

JIRA Issues: CSL-10537

Differential Revision: https://phabricator.twitter.biz/D600938