finagle-core: Add Address.ServiceFactory.apply Without Addr.Metadata
Problem

It's awkward ergonomically that `Address.apply` that takes a
`ServiceFactory` doesn't require `Addr.Metadata`, but the more specific
`Address.ServiceFactory.apply` does require `Addr.Metadata`. It's a rare
scenario where a user of the API would be supplying `Addr.Metadata` and
so it would be nicer if the `Address.ServiceFactory.apply` method did
not require it.

Solution

Add a variant of `Address.ServiceFactory.apply` which does not require
`Addr.Metadata` and defaults that value to `Addr.Metadata.empty`.

Differential Revision: https://phabricator.twitter.biz/D605438