finagle-core: Add Name.bound Variant Which Takes a Service
Problem

For testing Finagle clients, it's extremely useful to be able to pass it
a `Service` directly that the client can use instead of talking to a
server. This is possible because Finagle has a unified interface between
client and server where a server exports a `Service` and a client
imports it. Many long time Finagle users don't know that this is
possible though because to go from a `Service` to a `Name` (which is
necessary for `EndpointerStackClient.newService` et al.) requires
multiple hops from `Service` to `ServiceFactory` to `Address` to `Name`.

Solution

Provide an alternative to `Name.bound` which can take in a `Service`
directly and handles the necessary hops for the user.

Differential Revision: https://phabricator.twitter.biz/D605745