finagle-memcached: Turn Down Client Version Level
Problem

For services that contain a multitude of memcached clients, the logging
of whether or not Finagle is using the newer partitioned client or the
older non-partitioned client for memcached is quite noisy, as that
information is written out per client.

Solution

Lower the level on the messages which indicate which version of the
memcached client is being used. Because the logger is a
`java.util.logging.Logger`, these messages are now written at a 'fine'
level.

Result

Less noisy startup logs for services which have many memcached clients.

Differential Revision: https://phabricator.twitter.biz/D607487